// src/enums/race.ts
export const enum Race {
  INVALID = 0,
  BLOODELF = 1,
  DRAENEI = 2,
  DWARF = 3,
  GNOME = 4,
  GOBLIN = 5,
  HUMAN = 6,
  NIGHTELF = 7,
  ORC = 8,
  TAUREN = 9,
  TROLL = 10,
  UNDEAD = 11,
  WORGEN = 12,
  GOBLIN2 = 13,
  MURLOC = 14,
  DEMON = 15,
  SCOURGE = 16,
  MECHANICAL = 17,
  ELEMENTAL = 18,
  OGRE = 19,
  BEAST = 20,
  PET = 21,
  TOTEM = 22,
  PIRATE = 23,
  DRAGON = 24,
  BLANK = 25,
  ALL = 26,
  EGG = 38,
  NAGA = 92,
}